Move QueueState model helpers into a dedicated _queue_utils companion
Per PR review: instead of inlining the lower-level helpers directly in
_queue_state.py, keep them in a single _queue_utils.py companion module used
only by the OOP model. _queue_state.py now holds just the JobEntry/QueueState
classes; the model still does not depend on the procedural queue functions.
Mocks for the relocated helpers now target _queue_utils.
