+++ tests/dandi_compute_code/queue/model/conftest.py
@@ -44,7 +44,7 @@ def mock_dandi_assets_metadata() -> Iterator[None]:
     with (
         mock.patch("dandi_compute_code.queue._queue_state.load_assets_jsonld_metadata", return_value=empty_metadata),
         mock.patch(
-            "dandi_compute_code.queue._queue_state._load_upstream_assets_jsonld_metadata",
+            "dandi_compute_code.queue._queue_utils._load_upstream_assets_jsonld_metadata",
             return_value=empty_metadata,
         ),
     ):
